Move YourVoyages to own folder and refactor voyage display
